streamlit/7-ex-template-app-datascience-dashboard/requirements.txt
streamlit
streamlit-autorefresh
python-dotenv

gunicorn
pandas
numpy
matplotlib
seaborn
plotly

google-cloud-bigquery 
db-dtypes